Get project to render again.

+++ src/pellengine/components/sprite.h
@@ -7,7 +7,7 @@ namespace pellengine {
 
 struct Sprite {
   glm::vec4 color;
-  int zIndex;
+  float zIndex;
   bool dirty;
 };
